pyproject.toml
[build-system]
requires = [ "setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-mcp"
version = "0.1.0"
description = "Multi-account Gmail MCP server — manage 10+ Gmail inboxes through a single MCP server"
readme = "README.md"
keywords = [ "mcp", "gmail", "multi-account", "llm", "ai", "claude"]
license = "MIT"
requires-python = ">=3.10"
dependencies = [
 "fastapi>=0.115.12",
 "fastmcp>=3.1.1",
 "google-api-python-client>=2.168.0",
 "google-auth-httplib2>=0.2.0",
 "google-auth-oauthlib>=1.2.2",
 "httpx>=0.28.1",
 "py-key-value-aio>=0.3.0",
 "pyjwt>=2.12.0",
 "python-dotenv>=1.1.0",
 "pyyaml>=6.0.2",
 "cryptography>=45.0.0",
]

[project.scripts]
gmail-mcp = "main:main"

[tool.setuptools]
py-modules = ["main"]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["tests*", "docs*", "build", "dist"]

[tool.setuptools.package-data]
core = ["tool_tiers.yaml"]
